refactor: 分离视频信息操作组件
